requirements.txt
pandas
pip
python-dotenv
yfinance
streamlit
plotly
prophet
crewai
crewai_tools
